+++ tests/test_actions.py
@@ -273,5 +273,6 @@ async def test_default_action(
 
     channel = default_dispatcher_collecting.output_channel
     assert channel.messages == [
-        {u'text': u'default message', u'recipient_id': u'my-sender'}]
+        {u'text': u'sorry, I didn\'t get that, can you rephrase it?',
+         u'recipient_id': u'my-sender'}]
     assert events == [UserUtteranceReverted()]